+++ NEWS.md
@@ -5,6 +5,7 @@
 - Add `encode_images_in_messages()` to encode images in messages for `chat()` function.
 - Add `create_messages()` to create messages easily.
 - Helper functions for managing messages accept `...` parameter to pass additional options.
+- Update README and docs to reflect changes.
 
 # ollamar 1.2.0