+++ NEWS.md
@@ -1,4 +1,4 @@
-# ollamar (development version)
+# ollamar 1.2.1
 
 - `generate()` and `chat()` accept multiple images as prompts/messages.
 - Add functions to validate messages for `chat()` function: `validate_message()`, `validate_messages()`.